Moved route 'switcheventstatus/:id' from Events to Auth
